ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
insightface==0.7.3
onnxruntime==1.20.1
opencv-python-headless==4.10.0.82
numpy==1.26.4
psycopg2-binary==2.9.9
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.0
python-dotenv==1.0.1
